⚡️ re-add request caching
