+++ src/common/services/applemusic/index.ts
@@ -1,4 +1,5 @@
 import icon from '../../../../res/applemusic.svg'
+import { withCache } from '../../utils/cache'
 import { Resolvable, Searchable, Service } from '../types'
 import { resolve } from './resolve'
 import { search } from './search'
@@ -8,6 +9,6 @@ export const AppleMusic: Service & Searchable & Resolvable = {
   name: 'Apple Music',
   regex: /https?:\/\/music\.apple\.com\/(\w{2,4})\/album\/([^/]*)\/([^?]+)[^/]*/,
   icon,
-  search,
-  resolve,
+  search: withCache('applemusic', search),
+  resolve: withCache('applemusic', resolve),
 }
