+++ src/common/services/bandcamp/index.ts
@@ -1,4 +1,5 @@
 import icon from '../../../../res/bandcamp.svg'
+import { withCache } from '../../utils/cache'
 import { Embeddable, Resolvable, Searchable, Service } from '../types'
 import { embed } from './embed'
 import { resolve } from './resolve'
@@ -9,7 +10,7 @@ export const Bandcamp: Service & Searchable & Resolvable & Embeddable = {
   name: 'Bandcamp',
   regex: /https?:\/\/.*\.bandcamp\.com\/(track|album)\/.*/,
   icon,
-  search,
-  resolve,
-  embed,
+  search: withCache('bandcamp', search),
+  resolve: withCache('bandcamp', resolve),
+  embed: withCache('bandcamp', embed),
 }
