+++ src/common/services/discogs/index.ts
@@ -1,4 +1,5 @@
 import icon from '../../../../res/discogs.svg'
+import { withCache } from '../../utils/cache'
 import { Resolvable, Service } from '../types'
 import { regex } from './regex'
 import { resolve } from './resolve'
@@ -8,5 +9,5 @@ export const Discogs: Service & Resolvable = {
   name: 'Discogs',
   regex,
   icon,
-  resolve,
+  resolve: withCache('discogs', resolve),
 }
