+++ src/common/services/soundcloud/index.ts
@@ -1,4 +1,5 @@
 import icon from '../../../../res/soundcloud.svg'
+import { withCache } from '../../utils/cache'
 import { Embeddable, Resolvable, Searchable, Service } from '../types'
 import { embed } from './embed'
 import { resolve } from './resolve'
@@ -9,7 +10,7 @@ export const Soundcloud: Service & Searchable & Resolvable & Embeddable = {
   name: 'Soundcloud',
   regex: /((http:\/\/(soundcloud\.com\/.*|soundcloud\.com\/.*\/.*|soundcloud\.com\/.*\/sets\/.*|soundcloud\.com\/groups\/.*|snd\.sc\/.*))|(https:\/\/(soundcloud\.com\/.*|soundcloud\.com\/.*\/.*|soundcloud\.com\/.*\/sets\/.*|soundcloud\.com\/groups\/.*)))/,
   icon,
-  search,
-  resolve,
-  embed,
+  search: withCache('soundcloud', search),
+  resolve: withCache('soundcloud', resolve),
+  embed: withCache('soundcloud', embed),
 }
